User Resolver 분리
