+++ src/main/kotlin/com/undefined/platz/graphql/resolver/UserQueryResolver.kt
@@ -1,7 +1,6 @@
 package com.undefined.platz.graphql.resolver
 
 import com.expediagroup.graphql.scalars.ID
-import com.expediagroup.graphql.spring.operations.Mutation
 import com.expediagroup.graphql.spring.operations.Query
 import com.undefined.platz.graphql.model.User
 import com.undefined.platz.service.UserService
@@ -18,21 +17,5 @@ class UserQueryResolver(
 
     fun users(): List<User> {
         return userService.findAll().map { User(it) }
-    }
-}
-
-data class CreateUserInput(
-        var email: String,
-        val nickname: String,
-        val profilePicture: String
-)
-
-@Component
-class UserMutationResolver(
-        private val userService: UserService
-) : Mutation {
-    fun createUser(input: CreateUserInput): User {
-        val user = userService.create(input.email, input.nickname, input.profilePicture)
-        return User(user)
     }
 }
